StoryMakerGroups
** MIGRATE REQUIRED **
$python manage.py migrate

See project_template.models for StoryMakerGroup,  Shareable, and
SharerableBlogPost.

Shareable is an abstract class acting as a foundation for any object
that should enter the activity stream with sharing options. It handles
generating an appropriate action on save when possible.

Includes basic test of public and private stream permissions using
ShareableBlogPost.